[Cleanup] Cleanup in the traction control, antilock brake and engine code

+++ source/main/gameplay/CacheSystem.cpp
@@ -1247,7 +1247,7 @@ void CacheSystem::fillTruckDetailInfo(CacheEntry &entry, Ogre::DataStreamPtr ds,
 
 	if (r.engine)
 	{
-		entry.numgears   = (int)r.engine->getNumGears();
+		entry.numgears   = r.engine->getNumGears();
 		entry.minrpm     = r.engine->getMinRPM();
 		entry.maxrpm     = r.engine->getMaxRPM();
 		entry.torque     = r.engine->getEngineTorque();
